+++ packages/plugin-audio/web/src/audio-component.tsx
@@ -3,7 +3,13 @@ import AudioViewer from './audio-viewer';
 import type { AudioPluginEditorConfig } from './types';
 import { AUDIO_TYPE } from './types';
 import { DEFAULTS } from './defaults';
-import type { ComponentData, Helpers, RichContentTheme } from 'wix-rich-content-common';
+import type {
+  ComponentData,
+  Helpers,
+  RichContentTheme,
+  TranslationFunction,
+} from 'wix-rich-content-common';
+import { MediaItemErrorMsg } from 'wix-rich-content-ui-components';
 
 interface Props {
   componentData: ComponentData;
@@ -11,6 +17,7 @@ interface Props {
   theme: RichContentTheme;
   helpers: Helpers;
   isMobile: boolean;
+  t?: TranslationFunction;
   isLoading?: boolean;
   tempData?: boolean;
 }
@@ -19,17 +26,20 @@ class AudioComponent extends React.Component<Props> {
   static type = { AUDIO_TYPE };
 
   render() {
-    const { componentData, settings, theme, helpers, isMobile, isLoading } = this.props;
+    const { componentData, settings, theme, helpers, isMobile, isLoading, t } = this.props;
     return (
-      <AudioViewer
-        componentData={componentData}
-        settings={settings}
-        theme={theme}
-        helpers={helpers}
-        isMobile={isMobile}
-        isLoading={componentData?.tempData || isLoading}
-        disabled
-      />
+      <>
+        <AudioViewer
+          componentData={componentData}
+          settings={settings}
+          theme={theme}
+          helpers={helpers}
+          isMobile={isMobile}
+          isLoading={componentData?.tempData || isLoading}
+          disabled
+        />
+        {componentData?.error && <MediaItemErrorMsg error={componentData?.error} t={t} />}
+      </>
     );
   }
 }